services/api/requirements.txt
supabase==2.10.0
python-dotenv==1.0.0
requests>=2.31.0
tzdata>=2024.1
anthropic>=0.39.0
